requirements.txt
absl-py==0.7.1
astor==0.8.0
boto==2.49.0
boto3==1.9.219
botocore==1.12.219
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
docutils==0.15.2
emot==2.1
gast==0.2.2
gensim==3.8.0
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
jmespath==0.9.4
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.3
nltk==3.4.5
numpy==1.16.4
pandas==0.25.1
Pillow==6.1.0
pkg-resources==0.0.0
protobuf==3.9.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
requests==2.22.0
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
sklearn==0.0
smart-open==1.8.4
tb-nightly==1.14.0a20190603
tensorflow-gpu==2.0.0b1
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
torch==1.2.0+cu92
torchvision==0.4.0+cu92
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
